+++ src/components/Nav/ProjectSuggestion.svelte
@@ -1,0 +1,33 @@
+<script>
+  export let suggestion
+
+  const {
+    name,
+    ticker,
+    logoUrl = 'https://production-sanbase-images.s3.eu-central-1.amazonaws.com/sprite/default_project_logo.png',
+  } = suggestion
+</script>
+
+<template lang="pug">
+include /ui/mixins
+
+img(src='{logoUrl}', alt='Project logo', width='16', height='16')
+|{name}
+span ({ticker})
+
+</template>
+
+<style lang="scss">
+  @import '@/mixins';
+
+  span {
+    color: var(--waterloo);
+    margin-left: 3px;
+  }
+
+  img {
+    min-width: 16px;
+    min-height: 16px;
+    margin-right: 5px;
+  }
+</style>
